+++ xposed-module/run-jvm-tests.sh
@@ -20,6 +20,8 @@ OUT=.jvm-test-out; rm -rf "$OUT"; mkdir -p "$OUT"
     app/src/main/java/com/specter/module/gen/UsedStore.java \
     app/src/main/java/com/specter/module/gen/RootWriter.java \
     app/src/main/java/com/specter/module/gen/SessionMigrator.java \
+    app/src/main/java/com/specter/module/gen/WidevineL3.java \
+    app/src/main/java/com/specter/module/gen/GsfReset.java \
     app/src/main/java/com/specter/module/ui/DiagnosticsCmd.java \
     app/src/main/java/com/specter/module/ui/TraceParser.java \
     app/src/main/java/com/specter/module/ui/VaultChecksum.java \
@@ -30,6 +32,8 @@ OUT=.jvm-test-out; rm -rf "$OUT"; mkdir -p "$OUT"
     app/src/test/java/com/specter/module/gen/ProfileTest.java \
     app/src/test/java/com/specter/module/gen/RootWriterTest.java \
     app/src/test/java/com/specter/module/gen/SessionMigratorTest.java \
+    app/src/test/java/com/specter/module/gen/WidevineL3Test.java \
+    app/src/test/java/com/specter/module/gen/GsfResetTest.java \
     app/src/test/java/com/specter/module/ui/DiagnosticsCmdTest.java \
     app/src/test/java/com/specter/module/ui/TraceParserTest.java \
     app/src/test/java/com/specter/module/ui/VaultPortableTest.java \
@@ -40,6 +44,8 @@ OUT=.jvm-test-out; rm -rf "$OUT"; mkdir -p "$OUT"
 "$JAVA" -cp "$OUT" com.specter.module.gen.ProfileTest
 "$JAVA" -cp "$OUT" com.specter.module.gen.RootWriterTest
 "$JAVA" -cp "$OUT" com.specter.module.gen.SessionMigratorTest
+"$JAVA" -cp "$OUT" com.specter.module.gen.WidevineL3Test
+"$JAVA" -cp "$OUT" com.specter.module.gen.GsfResetTest
 "$JAVA" -cp "$OUT" com.specter.module.ui.DiagnosticsCmdTest
 "$JAVA" -cp "$OUT" com.specter.module.ui.TraceParserTest
 "$JAVA" -cp "$OUT" com.specter.module.ui.VaultPortableTest